+++ src/com/milanesachan/DRPGTest1/bot/handlers/Answerable.java
@@ -1,0 +1,8 @@
+package com.milanesachan.DRPGTest1.bot.handlers;
+
+public interface Answerable {
+
+    void answer(String answer);
+
+    void cancel();
+}