+++ src/com/milanesachan/DRPGTest1/bot/handlers/TesterHandler.java
@@ -6,7 +6,7 @@ import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 import java.awt.*;
 
-public class TesterHandler implements Handler{
+public class TesterHandler implements Handler, Answerable{
     private MessageChannel channel;
     private MessageReceivedEvent event;
 
@@ -17,15 +17,7 @@ public class TesterHandler implements Handler{
 
     @Override
     public void handle() {
-        String[] args = event.getMessage().getContentRaw().split(" ");
-        if(args.length>=3 && isInt(args[1]) && isInt(args[2])) {
-            int x = Integer.parseInt(args[1]);
-            int y = Integer.parseInt(args[2]);
-            Point p = new Point(x, y);
-            TestImageGenerator.getInstance().postImage(channel, event.getMember().getUser().getIdLong(), p);
-        }else{
-            channel.sendMessage("Incorrect format. Expected: '>test <x> <y>'").queue();
-        }
+        
     }
 
     private boolean isInt(String s){
@@ -35,5 +27,15 @@ public class TesterHandler implements Handler{
         }catch (NumberFormatException e){
             return false;
         }
+    }
+
+    @Override
+    public void answer(String answer) {
+
+    }
+
+    @Override
+    public void cancel() {
+
     }
 }