+++ changelog.md
@@ -6,10 +6,13 @@
 	- added statcards to the equipment
 	- unequip now removes your entire equipment
 	- created the 'Answerable' interface and the 'AnswerManager' for two-part commands
-	
+	- created the FusionHandler for the command '>fuse'
+	- added 'getStat(int)' to StatCard to get a stat determined from a number (0 = str, 1 = acc, 2 = vit, 3 = mag)
+
 - **Major**
 	- added 'stat cards' for the character to equip and have its statistics
 	- added "two-part commands" that will be used for complex interactions through commands
+	- added '>fuse' command to fuse two or more cards and get one, better card
 
 - **Fixed**